backend/requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
Pillow>=10.0.0
python-decouple>=3.8
django-filter>=24.2
whitenoise>=6.6.0
gunicorn>=21.2.0
